app/requirements.txt
ariadne==0.*
python-jose==3.*
passlib==1.*